templates/assistant/requirements.txt
# SPECTRA AI Assistant Requirements
# Production dependencies for AI assistant implementation

# Core dependencies
pyyaml>=6.0.1,<7.0
requests>=2.31.0,<3.0

# AI/ML dependencies  
openai>=1.0.0,<2.0

# Validation and testing
jsonschema>=4.17.0,<5.0

# Security and utilities
python-dateutil>=2.8.2,<3.0

# Development dependencies (comment out for production)
pytest>=7.4.0,<8.0
pytest-cov>=4.1.0,<5.0
pytest-mock>=3.11.0,<4.0
bandit>=1.7.5,<2.0
safety>=2.3.5,<3.0